requirements.txt
aiosqlite==0.20.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
fastapi==0.115.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
